prepareData Prepares data for comparison in 3 stages. 1. Match columns - filter dataframes to those columns that match and summarise differences 2. Match rows - filter dataframes to those rows that match and summarise differences 3. Coerce data

prepareData Prepares data for comparison in 3 stages. 1. Match columns - filter dataframes to those columns that match and summarise differences 2. Match rows - filter dataframes to those rows that match and summarise differences 3. Coerce data

prepareData(dfA, dfB, keys = NA, trimChars = TRUE)

Arguments

dfA
data frame. The first data object. dataCompareR will attempt to coerce all data objects to data frames.

dfB
data frame. The second data object. dataCompareR will attempt to coerce all data objects to data frames.

keys
String. Name of identifier column(s) used to compare dfA and dfB. NA if no identifier (row order will be used instead), a character for a single column name, or a vector of column names to match of multiple columns

trimChars
Boolean. If true, strings and factors have whitespace trimmed before comparison.

Value

dataCompareRObject containing details of the comparison

Examples

if (FALSE) dfA <- iris
if (FALSE) dfB <- iris
if (FALSE) keys <- NA
if (FALSE) prepareData(dfA,dfB,keys, trimChars = TRUE)