Save a report based on a dataCompareR object

Saves R markdown and HTML reports in the area specified by the user.

Uses knitr and markdown to create reports. Reports have the extensions .Rmd or .html. By default the table.css style sheet is used for format the html output.

saveReport(
  compareObject,
  reportName,
  reportLocation = ".",
  HTMLReport = TRUE,
  showInViewer = TRUE,
  stylesheet = NA,
  printAll = FALSE,
  ...
)

Arguments

compareObject
a dataCompareR object.

reportName
String. The name of the report. Reports will be saved as reportName.Rmd and (optionally) reportName.html in reportLocation

reportLocation
String. Location to save reports specified by the user. The R markdown and (optionally) HTML reports will be saved in this area

HTMLReport
Boolean. Option to output html report.

showInViewer
Boolean. Does the html report open automatically in the viewer?

stylesheet
String. Optional link to customised css stylesheet

printAll
Boolean. If TRUE, all mis-matches in the data are printed to the file. This acts as a shortcut to get all mismatches in the report, compared to passing the number in mismatchCount. When TRUE, overrides the mismatchCount field passed via ellipses

...
Optional arguments which will be passed to summary, for example mismatchCount

See also

Examples

 
if (FALSE) saveReport(rcObj, reportName = 'testReport')