Summarizing RCompare Output

Summarizing RCompare Output

# S3 method for dataCompareRobject
summary(object, mismatchCount = 5, ...)

Arguments

object
an dataCompareR object, usually a result of a call to rCompare.

mismatchCount
Integer. How many mismatches to include in tables

...
Passes any additional arguments (not used in current version)

Value

The function summary.dataCompareR computes and returns a list of summary details from the dataCompareR output given in object containing

datanameA

name of the first dataframe in the compare call

datanameB

name of the second dataframe in the compare call

nrowA

the number of rows in datanameA

nrowB

the number of rows in datanameB

version

the version of rCompare used to generate the dataCompareR object object

runtime

the date and time the dataCompareR object object was created

rversion

the version of R used

datasetSummary

a data frame containing the meta data information on datanameA and datanameB

ncolCommon

the number of columns of the same name contained in both datanameA and datanameB

ncolInAOnly

the number of columns only in datanameA

ncolInBOnly

the number of columns only in datanameB

ncolID

the number of columns used to match rows in datanameA and datanameB

typeMismatch

a data frame detailing which columns in both datanameA and datanameB have different class types

typeMismatchN

the number of columns with different variable types

nrowCommon

the number of rows with matching ID columns in both datanameA and datanameB

nrowInAOnly

the number of rows with non matching ID columns in datanameA

nrowInBOnly

the number of rows with non matching ID columns in datanameB

nrowSomeUnequal

the number of matched rows where at least one value is unequal

nrowAllEqual

the number of matched rows where all values are equal

ncolsAllEqual

the number of matched columns where all values are equal

ncolsSomeUnequal

the number of matched columns where at least one value is unequal

colsWithUnequalValues

a data frame detailing the mismatches for each matched column

nrowNAmisMatch

the number of matched numeric rows that contain a NA

maxDifference

the maximum difference between numeric columns from all matched columns

See also

Examples

rc1 <- rCompare(iris,iris) 

#> Running rCompare...

summary(rc1)

#> dataCompareR is generating the summary...

#> 
#> Data Comparison
#> ===============
#> 
#> Date comparison run: 2021-08-31 16:32:19  
#> Comparison run on R version 4.1.1 (2021-08-10)  
#> With dataCompareR version 0.1.3  
#> 
#> 
#> Meta Summary
#> ============
#> 
#> 
#> |Dataset Name |Number of Rows |Number of Columns |
#> |:------------|:--------------|:-----------------|
#> |iris         |150            |5                 |
#> |iris         |150            |5                 |
#> 
#> 
#> Variable Summary
#> ================
#> 
#> Number of columns in common: 5  
#> Number of columns only in iris: 0  
#> Number of columns only in iris: 0  
#> Number of columns with a type mismatch: 0  
#> No match key used, comparison is by row
#> 
#> 
#> 
#> Row Summary
#> ===========
#> 
#> Total number of rows read from iris: 150  
#> Total number of rows read from iris: 150    
#> Number of rows in common: 150  
#> Number of rows dropped from iris: 0  
#> Number of rows dropped from  iris: 0  
#> 
#> 
#> Data Values Comparison Summary
#> ==============================
#> 
#> Number of columns compared with ALL rows equal: 5  
#> Number of columns compared with SOME rows unequal: 0  
#> Number of columns with missing value differences: 0  
#> 
#> Columns with all rows equal : PETAL.LENGTH, PETAL.WIDTH, SEPAL.LENGTH, SEPAL.WIDTH, SPECIES
#>